Code cleanup: Cleaned up unused resources and reorganize codes
