--- src/app/demo/[name]/blocks/blank-page.tsx
@@ -1,9 +1,0 @@
-export default function BlankPage() {
-  return (
-    <div className="flex flex-col items-center justify-center gap-4 rounded-md py-96">
-      <div className="text-muted-foreground">
-        Start prompting to use your component!
-      </div>
-    </div>
-  );
-}
