--- src/app/demo/[name]/blocks/blank.tsx
@@ -1,8 +1,0 @@
-import BlankPage from "@/app/demo/[name]/blocks/blank-page";
-
-export const blank = {
-  name: "blank",
-  components: {
-    Default: <BlankPage />,
-  },
-};
