--- src/app/demo/[name]/blocks/shell-layout.tsx
@@ -1,23 +1,0 @@
-import React, { type ReactNode } from "react";
-
-import { BrandHeader } from "@/components/brand-header";
-import { BrandSidebar } from "@/components/brand-sidebar";
-import { SidebarProvider } from "@/components/ui/sidebar";
-import { Toaster } from "@/components/ui/sonner";
-
-export default function ShellLayout({
-  children,
-}: Readonly<{
-  children: ReactNode;
-}>) {
-  return (
-    <SidebarProvider>
-      <BrandHeader />
-      <BrandSidebar />
-      <main className="flex w-full justify-center">
-        <div className="container">{children}</div>
-      </main>
-      <Toaster />
-    </SidebarProvider>
-  );
-}
