--- src/app/demo/[name]/components/brand-header.tsx
@@ -1,13 +1,0 @@
-import { BrandHeader } from "@/components/brand-header";
-import { SidebarProvider } from "@/components/ui/sidebar";
-
-export const brandHeader = {
-  name: "brand-header",
-  components: {
-    Default: (
-      <SidebarProvider>
-        <BrandHeader />
-      </SidebarProvider>
-    ),
-  },
-};
