--- src/app/demo/[name]/components/brand-sidebar.tsx
@@ -1,13 +1,0 @@
-import { BrandSidebar } from "@/components/brand-sidebar";
-import { SidebarProvider } from "@/components/ui/sidebar";
-
-export const brandSidebar = {
-  name: "brand-sidebar",
-  components: {
-    Default: (
-      <SidebarProvider>
-        <BrandSidebar />
-      </SidebarProvider>
-    ),
-  },
-};
