--- src/app/demo/[name]/components/login.tsx
@@ -1,8 +1,0 @@
-import { Login } from "@/components/login";
-
-export const login = {
-  name: "login",
-  components: {
-    Default: <Login />,
-  },
-};
