--- src/app/demo/[name]/components/logo.tsx
@@ -1,8 +1,0 @@
-import { Logo } from "@/components/logo";
-
-export const logo = {
-  name: "logo",
-  components: {
-    Default: <Logo />,
-  },
-};
