+++ src/app/demo/[name]/components/top-bar.tsx
@@ -1,5 +1,5 @@
 import Topbar from "@/components/blocks/top-bar";
-import MinimalLayout from "./minimal-layout";
+import MinimalLayout from "../blocks/minimal-layout";
 
 export const topbar = {
   name: "topbar",
