+++ src/app/demo/[name]/index.tsx
@@ -1,19 +1,7 @@
 import type { ReactElement, ReactNode } from "react";
 
-// blocks
-import { blank } from "@/app/demo/[name]/blocks/blank";
-import { dashboard } from "@/app/demo/[name]/blocks/dashboard";
-import { store } from "@/app/demo/[name]/blocks/store";
-import { topbar } from "@/app/demo/[name]/blocks/top-bar";
-
 // components
-import { brandHeader } from "@/app/demo/[name]/components/brand-header";
-import { brandSidebar } from "@/app/demo/[name]/components/brand-sidebar";
-import { hero } from "@/app/demo/[name]/components/hero";
-import { login } from "@/app/demo/[name]/components/login";
-import { logo } from "@/app/demo/[name]/components/logo";
-import { productGrid } from "@/app/demo/[name]/components/product-grid";
-import { promo } from "@/app/demo/[name]/components/promo";
+import { topbar } from "@/app/demo/[name]/components/top-bar";
 
 // ui
 import { accordion } from "@/app/demo/[name]/ui/accordion";
@@ -26,14 +14,14 @@ import { breadcrumb } from "@/app/demo/[name]/ui/breadcrumb";
 import { button } from "@/app/demo/[name]/ui/button";
 import { calendar } from "@/app/demo/[name]/ui/calendar";
 import { card } from "@/app/demo/[name]/ui/card";
-import {carousel} from '@/app/demo/[name]/ui/carousel';
+import { carousel } from "@/app/demo/[name]/ui/carousel";
 import { chart } from "@/app/demo/[name]/ui/chart";
 import { checkbox } from "@/app/demo/[name]/ui/checkbox";
 import { contextMenu } from "@/app/demo/[name]/ui/context-menu";
 import { dataTable } from "@/app/demo/[name]/ui/data-table";
 import { datePicker } from "@/app/demo/[name]/ui/date-picker";
 import { dialog } from "@/app/demo/[name]/ui/dialog";
-import {drawer} from "@/app/demo/[name]/ui/drawer";
+import { drawer } from "@/app/demo/[name]/ui/drawer";
 import { dropdownMenu } from "@/app/demo/[name]/ui/dropdown-menu";
 import { input } from "@/app/demo/[name]/ui/input";
 import { inputOtp } from "@/app/demo/[name]/ui/inputOtp";
@@ -44,7 +32,7 @@ import { radioGroup } from "@/app/demo/[name]/ui/radio-group";
 import { resizable } from "@/app/demo/[name]/ui/resizable";
 import { select } from "@/app/demo/[name]/ui/select";
 import { separator } from "@/app/demo/[name]/ui/separator";
-import {sheet} from "@/app/demo/[name]/ui/sheet";
+import { sheet } from "@/app/demo/[name]/ui/sheet";
 import { skeleton } from "@/app/demo/[name]/ui/skeleton";
 import { slider } from "@/app/demo/[name]/ui/slider";
 import { sonner } from "@/app/demo/[name]/ui/sonner";
@@ -72,24 +60,12 @@ interface Demo {
 }
 
 export const demos: { [name: string]: Demo } = {
-  // blocks
-  blank,
-  store,
-  dashboard,
-  topbar,
-
   // components
-  hero,
-  login,
-  promo,
-  logo,
-  "brand-header": brandHeader,
-  "brand-sidebar": brandSidebar,
-  "product-grid": productGrid,
+  topbar,
 
   // ui
   accordion,
-  "alert-dialog":alertDialog,
+  "alert-dialog": alertDialog,
   alert,
   "aspect-ratio": aspectRatio,
   avatar,
@@ -108,7 +84,7 @@ export const demos: { [name: string]: Demo } = {
   dialog,
   "date-picker": datePicker,
   "data-table": dataTable,
-   drawer,
+  drawer,
   "dropdown-menu": dropdownMenu,
   input,
   inputOtp,
@@ -116,7 +92,7 @@ export const demos: { [name: string]: Demo } = {
   // "menu-bar": menuBar,
   "navigation-menu": navigationMenu,
   progress,
-  "scroll-area" :scrollArea,
+  "scroll-area": scrollArea,
   "radio-group": radioGroup,
   resizable,
   select,
